model h/cpp renamed into reconstruction h/cpp
reconstruction h/cpp is more consistent with class name
