+++ src/processor/incremental_triangulator.h
@@ -2,7 +2,7 @@
 #define INC_3D_RECONSTRUCTION_INCREMENTAL_TRIANGULATOR_H
 
 #include "../storage.h"
-#include "../model.h"
+#include "../reconstruction.h"
 #include "../optimization.h"
 #include "../refinement.h"
